backend/requirements.txt
flask
psycopg2-binary
flask-cors
flask-sqlalchemy
python-dotenv